refactor: 쿼리 생성 StringBuilder -> String.format 변경
