+++ src/main/java/persistence/sql/Table.java
@@ -1,0 +1,19 @@
+package persistence.sql;
+
+public class Table {
+
+    public String name(Class<?> clz) {
+        final jakarta.persistence.Table tableAnnotation = clz.getAnnotation(jakarta.persistence.Table.class);
+
+        if (existTableNameProperty(tableAnnotation)) {
+            return tableAnnotation.name().toLowerCase();
+        }
+
+        return clz.getSimpleName().toLowerCase();
+    }
+
+    private static boolean existTableNameProperty(jakarta.persistence.Table tableAnnotation) {
+        return tableAnnotation != null && !tableAnnotation.name().isEmpty();
+    }
+
+}
